+++ src/extension.ts
@@ -4,11 +4,13 @@ import * as vscode from 'vscode';
 
 import gekkoServerInfoHandler from './handler/gekkoServerInfo.handler';
 import backtestActiveEditor from './handler/backtestActiveEditor.handler';
+import createWorkspaceHandler from './handler/createWorkspace.handler';
 
 export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerCommand('extension.gekkoServerInfo', gekkoServerInfoHandler));
 	context.subscriptions.push(vscode.commands.registerCommand('extension.backtestStrategy', backtestActiveEditor));
+	context.subscriptions.push(vscode.commands.registerCommand('extension.createWorkspace', createWorkspaceHandler));
 }
 
 // this method is called when your extension is deactivated