+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Create a CSV with all performance reports and add a new line for each new backtest executed using this plugin.
 * Intelissence based on Gekko core. Use the vscode api to help to indicate the essentials functions provided by the gekko since its not exposed by one library.
 
+## [0.2.0]
+
+* Create Workspace command.
+* Support to `vscode-gekko-ext.json` under a Workspace.
+
 ## [0.1.0]
 
 * Use the `backtest.json` under a strategy package to retrieve the a configuration of the backtest.